Cleanup of some database items.
* No longer validating collecion names as there was no way to allow people to override.  Could be rethought in the future.
