+++ setup.cfg
@@ -61,6 +61,7 @@ install_requires =
     ruamel.yaml
     scalpl
     scipy
+    tqdm
 # The usage of test_requires is discouraged, see `Dependency Management` docs
 # tests_require = pytest; pytest-cov
 # Require a specific Python version, e.g. Python 2.7 or >= 3.4
@@ -173,18 +174,18 @@ source =
 [coverage:report]
 # Regexes for lines to exclude from consideration
 exclude_lines =
-    # Have to re-enable the standard pragma
+# Have to re-enable the standard pragma
     pragma: no cover
 
-    # Don't complain about missing debug-only code:
+# Don't complain about missing debug-only code:
     def __repr__
     if self\.debug
 
-    # Don't complain if tests don't hit defensive assertion code:
+# Don't complain if tests don't hit defensive assertion code:
     raise AssertionError
     raise NotImplementedError
 
-    # Don't complain if non-runnable code isn't run:
+# Don't complain if non-runnable code isn't run:
     if 0:
     if __name__ == .__main__.:
